# backend/swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Tarot API Test",
    "description": "오늘도 화이팅",
    "version": "1.0.0"
  },
  "host": "43.202.208.226:3001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "JWT Access Token 입력 : 'Bearer {token}'"
    }
  },
  "paths": {
...
    "/result/stream": {
      "post": {
        "tags": [
          "Result"
        ],
        "summary": "타로 결과 GPT 요청",
        "description": "타로 결과를 API에 요청하고 결과를 반환함",
        "parameters": [
          {
            "name": "cards",
            "in": "query",
            "description": "카드 배열",
            "required": true,
            "type": "array",
            "example": "[Ace of Wands, Ace of Cups, Ace of Swords]",
            "schema": {
              "type": "object",
              "properties": {
                "cards": {
                  "type": "string",
                  "example": "[Ace of Wands, Ace of Cups, Ace of Swords]"
                }
              }
            }
          },
          {
            "name": "ask",
            "in": "query",
            "description": "질문",
            "required": true,
            "type": "string",
            "example": "오늘의 운세는 어떤가요?",
            "schema": {
              "type": "object",
              "properties": {
                "ask": {
                  "type": "string",
                  "example": "오늘의 운세는 어떤가요?"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "GPT API 요청 성공",
            "schema": {
              "type": "object",
              "properties": {
                "json": {
                  "type": "object",
                  "properties": {
                    "card1": {
                      "type": "string",
                      "example": "Ace of Wands"
                    },
                    "card2": {
                      "type": "string",
                      "example": "Ace of Cups"
                    },
                    "card3": {
                      "type": "string",
                      "example": "Ace of Swords"
                    },
                    "ask": {
                      "type": "string",
                      "example": "오늘의 운세는 어떤가요?"
                    },
                    "answer": {
                      "type": "string",
                      "example": "오늘은 행운이 가득할 거예요!"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "유효하지않은 데이터",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "데이터가 유효하지 않습니다. (널값, 누락 등)"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "GPT API 요청 중 오류 발생",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "GPT에서 오류가 발생해 데이터를 불러올 수 없습니다!"
                },
                "error": {
                  "type": "string",
                  "example": ""
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/result/save": {
      "post": {
        "tags": [
          "Result"
        ],
        "summary": "종합 결과 저장",
        "description": "사용자의 질문과 관련된 타로 종합 결과 및 뽑은 카드 정보를 저장합니다. 사용자의 운(luck)에 따라 카드의 수가 결정됩니다.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "타로 결과 및 뽑은 카드 정보 저장을 위한 요청값",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "poll_id": {
                  "type": "string",
                  "example": "example"
                },
                "question": {
                  "type": "string",
                  "example": "사용자 질문"
                },
                "result_explanation": {
                  "type": "string",
                  "example": "종합 결과"
                },
                "master_name": {
                  "type": "string",
                  "example": "마스터 이름"
                },
                "luck": {
                  "type": "string",
                  "example": "test_luck"
                },
                "cards": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "card_image_url": {
                        "type": "string",
                        "example": "url3"
                      },
                      "card_explanation": {
                        "type": "string",
                        "example": "explanation3"
                      },
                      "card_eng_name": {
                        "type": "string",
                        "example": "eng3"
                      },
                      "card_kor_name": {
                        "type": "string",
                        "example": "kor3"
                      }
                    }
                  }
                }
              },
              "required": [
                "poll_id",
                "question",
                "result_explanation",
                "master_name",
                "luck",
                "cards"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "타로 종합 결과와 카드 정보가 성공적으로 저장됨",
            "schema": {
              "$ref": "#/definitions/SaveResponse"
            }
          },
          "400": {
            "description": "잘못된 요청, 필수 파라미터 누락"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        }
      }
    },
...
  }
}